+++ examples/resources/onepasswordorg_group/resource.tf
@@ -1,4 +1,4 @@
-resource "onepasswordorg_group" "test_group" {
+resource "onepasswordorg_group" "test" {
   name        = "test-group"
   description = "Group for testing"
 }